Move test case methods into a trait
