+++ ProphecyTrait.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Prophecy\PhpUnit;
+
+use Prophecy\Exception\Prediction\PredictionException;
+use Prophecy\Prophet;
+
+trait ProphecyTrait
+{
+    /**
+     * @var Prophet
+     */
+    private $prophet;
+
+    /**
+     * @param string|null $classOrInterface
+     * @return \Prophecy\Prophecy\ObjectProphecy
+     * @throws \LogicException
+     */
+    protected function prophesize($classOrInterface = null)
+    {
+        if (null === $this->prophet) {
+            throw new \LogicException(sprintf('The setUp method of %s must be called to initialize Prophecy.', __CLASS__));
+        }
+
+        return $this->prophet->prophesize($classOrInterface);
+    }
+
+    protected function setUp()
+    {
+        $this->prophet = new Prophet();
+    }
+
+    protected function assertPostConditions()
+    {
+        $this->prophet->checkPredictions();
+    }
+
+    protected function tearDown()
+    {
+        $this->prophet = null;
+    }
+
+    protected function onNotSuccessfulTest(\Exception $e)
+    {
+        if ($e instanceof PredictionException) {
+            $e = new \PHPUnit_Framework_AssertionFailedError($e->getMessage(), $e->getCode(), $e);
+        }
+
+        return parent::onNotSuccessfulTest($e);
+    }
+}
